Made appropriate changes for VehicleControl to work with AnvelBehavior
